fixing highlight behavior
I'm deciding not to highlight objects/arrows while dragging. Objects that are inferred from arrows are not stored in state, so there is no way to store whether they are being edited, other than to consulting drag state. Consulting drag state however gets into messy logic where we're checking a polymorphic drag state object to see if it is of a certain type. A proper approach might add a generic method to all drag classes, but this would introduce unnecessary complexity that would only serve a very narrow application. This is why I've decided to remove highlights from objects while dragging, which still appears to be very sensible behavior.