consolidating Diagram and DiagramViewState
The distinction has proven not to be useful, operations have either operated on lists of arrows or objects, or the coupled diagram+view, never diagrams or views individually. Meanwhile, the lack of a coupled diagram+view concept has complicated method signatures for commands and prevented the move towards separating command input and output.

Operations that need the entire view state are not seen as likely. Operations may exist exclusively on diagrams, but if an operation were needed, it could operate on the coupled diagram+view as an extension of the exclusive "Diagram" object as an interface.