commands now return diagram state directly
Diagram objects are now stored as records within undo/redo history. This simplifies command implementation vastly since now only the forward function needs to be implemented. The memory footprint of diagrams remains small in the undo/redo history, since old state still often needed to be stored, and only references need to be stored in the event no change occurs. This is reliable so long as diagram references are reused if and only if no change occurs. This practice has a side benefit, in that it also allows views to performantly check for updates, since referential equality is now a guaranteed indicator that no change has occurred.